testing and updating part 1 functions.

+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: nhariman <marvin@codam.nl>                   +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/11/01 17:06:48 by nhariman       #+#    #+#                */
-/*   Updated: 2019/11/01 18:30:07 by nhariman      ########   odam.nl         */
+/*   Updated: 2019/11/03 15:23:35 by nhariman      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
 	unsigned char	*str1;
 	unsigned char	*str2;
-	int				i;
+	size_t			i;
 
 	str1 = (unsigned char *)s1;
 	str2 = (unsigned char *)s2;
